feat: change util and update Box component
